requirements.txt
joblib==1.5.3
matplotlib==3.10.8
numpy==2.4.0
pandas==2.3.3
plotly==6.5.0
psutil==7.1.3
pytest==9.0.2
PyYAML==6.0.3
rich==14.2.0
scapy==2.6.1
scikit_learn==1.8.0
seaborn==0.13.2
streamlit==1.52.2
tensorflow==2.20.0
